feat: support create/update time in asset metadata
The create and update time fields in the Asset proto are meant to track
when the asset was first created or updated which would correspond to
the first and last time the asset was ingested.

Therefore, add create/update time fields inside specific asset metadata
protos:
- Bucket
- Dashboard
- Job
- Table
- Topic
- User

This will track, for example, when a table was created in the database
and the last time the table structure was updated.